LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-with-step.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<input type=date id=date value="2011-05-01" step="9">

<script>
description('Test if step limits available dates in calendar picker.');

openPicker(document.getElementById('date'), test1);

function test1() {
    debug("Step without min.");
    shouldBe('availableDatesInCurrentMonth().join(",")', '["2011-05-08", "2011-05-17", "2011-05-26"].join(",")');

    // Close popup
    eventSender.keyDown('\x1B');
    waitUntilClosing(test1AfterClose);
}

function test1AfterClose() {
    document.getElementById('date').min = "2011-05-01";
    
    // Reopen popup
    openPicker(document.getElementById('date'), test2);
}

function test2() {
    debug("Step when min is set.");
    shouldBe('availableDatesInCurrentMonth().join(",")', '["2011-05-01", "2011-05-10", "2011-05-19", "2011-05-28"].join(",")');
    finishJSTest();
}

function availableDatesInCurrentMonth() {
    return Array.prototype.map.call(popupWindow.document.querySelectorAll(".day.available:not(.not-this-month)"), function(element) {
        return element.dataset.submitValue;
    }).sort();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
